# Conflicts:
#	Modules/LoggerService/Sources/LoggerService/Private/NetLogger.cpp


Former-commit-id: 84e564f0747284c63271398d143cb81e683dc69f

+++ Sources/Internal/FileSystem/DynamicMemoryFile.cpp
@@ -112,7 +112,7 @@ uint64 DynamicMemoryFile::GetPos() const
 
 uint64 DynamicMemoryFile::GetSize() const
 {
-    return static_cast<uint32>(data.size());
+    return static_cast<uint64>(data.size());
 }
 
 bool DynamicMemoryFile::Seek(int64 position, eFileSeek seekType)
